+++ CHANGES.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+* Add support for multiple values of the same attribute (#95)
+
 * Add experimental async client and server implementation for python >=3.5.
 
 * Add IPv6 bind support for client and server.